chore: update naming for tagWrapper for consistency
