+++ src/cpu.ts
@@ -1,10 +1,14 @@
 /* eslint-disable @typescript-eslint/explicit-module-boundary-types */
-import { CpuProfiler } from '@datadog/pprof'
+import { CpuProfiler, encode } from '@datadog/pprof'
+import { perftools } from '@datadog/pprof/proto/profile'
 import debug from 'debug'
+
+// eslint-disable-next-line @typescript-eslint/no-explicit-any
+type ShamefulAny = any
 
 const log = debug('pyroscope::cpu')
 
-import { checkConfigured, uploadProfile } from './index'
+import { checkConfigured, config, processProfile, uploadProfile } from './index'
 
 const cpuProfiler = new CpuProfiler()
 
@@ -51,7 +55,7 @@ export function setCpuLabels(labels: Record<string, unknown>) {
   cpuProfiler.labels = labels
 }
 
-export function getCpuLabels(): any {
+export function getCpuLabels(): unknown {
   return cpuProfiler.labels
 }
 
@@ -59,13 +63,38 @@ export function tag(key: string, value: number | string | undefined) {
   cpuProfiler.labels = { ...cpuProfiler.labels, [key]: value }
 }
 
-export function tagCall(
+
+export function collectCpu(seconds: number): Promise<Buffer> {
+  if (!config.configured) {
+    throw 'Pyroscope is not configured. Please call init() first.'
+  }
+  log('Pyroscope has started CPU Profiling')
+  return new Promise((resolve, reject) => {
+    cpuProfiler.start(100);
+
+    setInterval(() => {
+      log('Collecting cpu profile')
+      const profile = cpuProfiler.profile();
+      if (profile) {
+        log('Cpu profile collected')
+        const newProfile = processProfile(profile);
+        if ( newProfile ) {
+          encode(newProfile).then(profile => resolve(profile))
+        }
+      } else {
+        log('Cpu profile collection failed')
+      }
+    }, seconds * 1000);
+  }
+}
+
+export function tagWrapper(
   key: string,
-  value: number | string | undefined,
+  value: string | undefined,
   fn: () => void,
   ...args: unknown[]
 ) {
   tag(key, value)
-  ;(fn as any)(...args)
+  ;(fn as ShamefulAny)(...args)
   tag(key, undefined)
 }
