+++ src/index.ts
@@ -7,7 +7,7 @@ import FormData from 'form-data'
 
 type TagList = Record<string, any>
 
-const log = debug('pyroscope')
+export const log = debug('pyroscope')
 
 const cloudHostnameSuffix = 'pyroscope.cloud'
 
@@ -21,11 +21,10 @@ export interface PyroscopeConfig {
   configured: boolean
 }
 
-const INTERVAL = 10000
-const SAMPLERATE = 100
+export const INTERVAL = 10000
+export const SAMPLERATE = 100
 
-export
-const config: PyroscopeConfig = {
+export const config: PyroscopeConfig = {
   serverAddress: process.env['PYROSCOPE_SERVER_ADDRESS'],
   appName: process.env['PYROSCOPE_APPLICATION_NAME'] || '',
   sm: undefined,
@@ -179,52 +178,7 @@ export async function uploadProfile(
   }
 }
 
-// Could be false or a function to stop heap profiling
-let heapProfilingTimer: undefined | NodeJS.Timer = undefined
-let isWallProfilingRunning = false
-
-export async function collectCpu(seconds?: number): Promise<Buffer> {
-  if (!config.configured) {
-    throw 'Pyroscope is not configured. Please call init() first.'
-  }
-
-  try {
-    const profile = await pprof.time.profile({
-      lineNumbers: true,
-      sourceMapper: config.sm,
-      durationMillis: (seconds || 10) * 1000 || INTERVAL,
-      intervalMicros: 10000,
-    })
-
-    const newProfile = processProfile(profile)
-    if (newProfile) {
-      return pprof.encode(newProfile)
-    } else {
-      return Buffer.from('', 'utf8')
-    }
-  } catch (e) {
-    log(e)
-    return Buffer.from('', 'utf8')
-  }
-}
-
-export async function collectHeap(): Promise<Buffer> {
-  if (!config.configured) {
-    throw 'Pyroscope is not configured. Please call init() first.'
-  }
-
-  log('Collecting heap...')
-  const profile = pprof.heap.profile(undefined, config.sm)
-  const newProfile = processProfile(profile)
-  if (newProfile) {
-    return pprof.encode(newProfile)
-  } else {
-    return Buffer.from('', 'utf8')
-  }
-}
-
-export
-function checkConfigured() {
+export function checkConfigured() {
   if (!config.configured) {
     throw 'Pyroscope is not configured. Please call init() first.'
   }
@@ -238,44 +192,6 @@ function checkConfigured() {
   }
 }
 
-export function startWallProfiling(): void {
-  checkConfigured()
-
-  log('Pyroscope has started CPU Profiling')
-  isWallProfilingRunning = true
-
-  const profilingRound = () => {
-    log('Collecting CPU Profile')
-    pprof.time
-      .profile({
-        lineNumbers: true,
-        sourceMapper: config.sm,
-        durationMillis: INTERVAL,
-        intervalMicros: 10000,
-      })
-      .then((profile) => {
-        log('CPU Profile collected')
-        if (isWallProfilingRunning) {
-          setImmediate(profilingRound)
-        }
-        log('CPU Profile uploading')
-        return uploadProfile(profile)
-      })
-      .then((d) => {
-        log('CPU Profile has been uploaded')
-      })
-      .catch((e) => {
-        log(e)
-      })
-  }
-  profilingRound()
-}
-
-// It doesn't stop it immediately, just wait until it ends
-export function stopWallProfiling(): void {
-  isWallProfilingRunning = false
-}
-
 export function start(): void {
   startCpuProfiling()
   startHeapProfiling()
@@ -286,74 +202,38 @@ export function stop(): void {
   stopHeapProfiling()
 }
 
-let isHeapCollectingStarted = false
-
-export function startHeapCollecting() {
-  if (!config.configured) {
-    throw 'Pyroscope is not configured. Please call init() first.'
-  }
-
-  if (isHeapCollectingStarted) {
-    log('Heap collecting is already started')
-    return
-  }
-
-  const intervalBytes = 1024 * 512
-  const stackDepth = 32
-
-  log('Pyroscope has started heap profiling')
-
-  pprof.heap.start(intervalBytes, stackDepth)
-  isHeapCollectingStarted = true
-}
-
-export function startHeapProfiling(): void {
-  checkConfigured()
-
-  if (heapProfilingTimer) {
-    log('Pyroscope has already started heap profiling')
-    return
-  }
-
-  startHeapCollecting()
-
-  heapProfilingTimer = setInterval(() => {
-    log('Collecting heap profile')
-    const profile = pprof.heap.profile(undefined, config.sm)
-    log('Heap profile collected...')
-    uploadProfile(profile).then(() => log('Heap profile uploaded...'))
-  }, INTERVAL)
-}
-
-export function stopHeapCollecting() {
-  pprof.heap.stop()
-  isHeapCollectingStarted = false
-}
-
-export function stopHeapProfiling(): void {
-  if (heapProfilingTimer) {
-    log('Stopping heap profiling')
-    clearInterval(heapProfilingTimer)
-    heapProfilingTimer = undefined
-    stopHeapCollecting()
-  }
-}
-
 import {
   startCpuProfiling,
   stopCpuProfiling,
   setCpuLabels,
   getCpuLabels,
-  tagCall,
+  tagWrapper,
   tag,
+  collectCpu,
 } from './cpu.js'
 export {
   startCpuProfiling,
   stopCpuProfiling,
   setCpuLabels,
   getCpuLabels,
-  tagCall,
+  collectCpu,
+  tagWrapper,
   tag,
+}
+import {
+  startHeapProfiling,
+  stopHeapProfiling,
+  collectHeap,
+  startHeapCollecting,
+  stopHeapCollecting,
+} from './heap.js'
+import { startWallProfiling, stopWallProfiling, collectWall } from './wall.js'
+export {
+  startHeapProfiling,
+  stopHeapProfiling,
+  collectHeap,
+  startHeapCollecting,
+  stopHeapCollecting,
 }
 import expressMiddleware from './express.js'
 export { expressMiddleware }
@@ -367,7 +247,7 @@ export default {
   startHeapProfiling,
   stopHeapProfiling,
   collectCpu,
-  collectWall: collectCpu,
+  collectWall,
   collectHeap,
   startHeapCollecting,
   stopHeapCollecting,
